lab_2_star/app_requirements.txt
flask==3.1.2
mysql-connector-python==8.1.0